[INTERNAL] removed dead code
The code for selecting projects that a user wanted to share got orphaned
with the introduction of the partial sharing feature which allowed to
select concrete resources instead of projects only.

I am just wondering why the code was not deleted and existed for more
than 2 years now doing absolutely nothing.


Change-Id: I1a32972e02a993d147c536f16224bc1e474e23a8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1163
Reviewed-by: Holger Freyther <holger@freyther.de>
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>