[FIX] Jupiter documents were registered to early for clients 2
Follow-up-patch for commit c093321 (see
http://saros-build.imp.fu-berlin.de/gerrit/#/c/1126/3).

Fixes adding Projects later on in the session, by checking the
sarosSession.userHasProjects() method when adding clients to a new
JupiterDocumentServer.

This solution coincidentally works for partially shared projects as the
toSPath()-method only checks that the project exists and not the
resource itself. Therefore the Jupiter-Activities for partially shared
resources will reach the ConcurrenctDocumentClient even if the resource
itself has not yet been received and update the Jupiter-Timestamp
correctly. (Only the EditorManager realizes that the resource is not
present and therefore doesn't apply it silently). But this is not a very
good solution.

TODO: Implement a clear solution for partially shared projects, by
tracking the projects AND resources, that are already shared for every
user individually.

Change-Id: I5e478b6b7bf255bce8798c395eff0eb326b1cccf
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1171
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>