[FIX] ensure that all packet extensions uses an alias
You would be surprised what will happen if you refactor the class name
or move the class to another package: It will break compatibility for no
reason.

I was also very surprised that this change discovered a bug where I
installed a wrong packet filter in the IPN but somehow it had no
impact because I used the same XML element twice (C&P is evil :P) for
different extensions.

Change-Id: I2a8e1ea0424c33bb3a45880b33036b01e7d2416d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1202
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>