[INTERNAL] removed duplicated (un)marshalling
The FileList class offered methods to perform marshalling. Seems to be a
left over and is no longer needed as the marshalling and unmarshalling
is correctly performed as long as the class is passed into the provider
who is responsible for the extensions that include this class.

Also removed the compression as the networklayer will take care of it if
the extension is send by bytestream. As for XMPP messages the connection
should be responsible to enable XMPP stream compression during
connection negotiation with the XMPP server.

Change-Id: I072efafb9bfc67e629abc1afb10b0240caf654be
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1206
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>