[GUI] inconsistency icon now "flashes" on detected inconsistency
I am pretty sure this will violate the default Eclipse UX design guide
as this change turns the Saros View much more into a GUI of a computer
game but: if you want the user to pay attention of something you must
catch its attention ... and making something flashing will mostly not be
left unrecognized.

If this is too disturbing we can still remove it.

Change-Id: I8391aba743fd484dd3a74d0ab9b3b350387bf7c6
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1208
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>