[INTERNAL] ensure the SelectionService will not throw NPE
Depending on the current thread the selection service will not be
available. As this service is almost only used in the actions you can
find a very ugly pattern that was C&P throughout all actions.

try
{
  Selection* = Selection*.get*
}
catch(NullPointerException)

The NPE was caused due an invalid thread access because not ALL !
listeners from the Saros Logic are invoked in the SWT Thread context.

TODO: Refactor all actions but note: even in the SWT Thread context the
selection service might be not available e.g the workbench is shutting
down.

Also removed a marker interface as it was never queried and thus not
needed at all.

REFACTOR: replaced generic type identifiers with T and S to be conform
with the Java coding conventions regarding generic type name parameter
naming.

Change-Id: I736d078e9e7c0320708491d190474d54fce0963f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1216
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>