[INTERNAL] added AWT UISynchronizer for multi-platform support
Just read 'Portierung von Saros auf andere Entwicklungsumgebungen on the
thesis Website'. Well Netbeans uses AWT.

But to be honest, this is not a topic for a Bachelor Thesis. This is not
even a topic for just only one D/M Thesis.

Starting to split components depending on their
implementation and to fetch them to build a product is a fully qualified
paid job (production team, btw we currently use Ant4->ECLIPSE<- for
production) along with proper SCM layout (this class
is definitely misplaced here in that package) and last not least
qualified developers using proper software engineering
practices.

If you would really like to support multi platforms:

Create a basic core that does not need to be compiled against IDE
specific stuff (a.k.a compile against interfaces).

Compile the specific implementation against the interfaces.

Make sure you setup the context (Pico Container) properly. As this may
be hard to include this during compile time it may have to be setup
during runtime using meta-programming (config files)

And btw ... the F5 button will no longer be the "compile button". It may
be much harder to correctly checkout the needed components. I do not
even know if GIT is capable of this compared to the "easy" use of config
specs in Clearcase or Perforce equivalents.

Change-Id: Iddb15b59d50083b1cb2e8ac46c8a979b8aeed6d2
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1218
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>