[FIX] ensure Actions are disposed
It is up to Eclipse to decide whether to hide or to dispose a view part.

In an old stage of Saros Actions were handled by the Pico container.
Somehow the developer who actually changes the behavior forget to ensure
that those actions must now be *disposed* because they are now
may be created more than once during the plugin life time.



Change-Id: Ibb1031cd4292c925fef13db25db629b1fa1351bd
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1224
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>