[GUI] removed the *flashing* of the taskbar
If you looking very very VERY carefully at the call that is made even an
***** will notice that this has nothing to do with flashing the task
bar.

Flashing the taskbar is OS specific and it neither supported by SWT nor
AWT / SWING. The reason while the taskbar was flashing is that most
newer Windows OS prevents focus stealing and just inform the user by
flashing the taskbar indicating that the applications wants to receive
the focus.

Maybe this will work fine on Window OS but this may fail badly on *nix
systems where the WindowManager will not forbid those calls and will
result in a very bad UX (somebody remember the *good* old days back in
1990-2000 with auto pop-ups in the IE ? finally you give up and push the
ON/OFF button of your PC).

See also:
http://msdn.microsoft.com/en-us/library/windows/desktop/ms633539%28v=vs.85%29.aspx
http://msdn.microsoft.com/en-us/library/windows/desktop/ms679347%28v=vs.85%29.aspx

Again: we may support this on Windows Platforms via JNA ... but well yes
this will only work for Windows.

Real solution: write native C code that compiles to DLL or SO files.
Problem: Must be compiled for every OS and every architecture (0x86,
0x64 etc.) along against correct A/W Win32 API. Must be loaded during
runtime.

Last but not least to compile C code on Win32 we must use MiniGW which
is much harder to setup than VisualStudio where we have to pay attention
if can distribute the compiled code with the current academic license
that is available to use due to MSDNAA. And I am pretty sure nobody
likes to compile C code thus we have to include the compiled library
into the SCM which is odd. A good example ... the Xuggler binaries ...
needed for screensharing ... are not available anymore for Win32. Do you
really expect the user to compile a library (requires download of
VisualStudio) to just use the experimental screen-sharing in Saros ?!

Java is Java, Eclipse used SWT and SWT is just support a small set of
possibilities the OS is offering because it has to work on EVERY
platform.

Conclusion: Java is not the platform to use when you need special
features that you will encounter in many native programs to display
*awareness* information because we (will never) have not access to
those function. Accept it or switch the language + destination platform
of your choice.

Change-Id: I52a6b11567b35dc3a67e270e1b64765567570e7c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1229
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>