[INTERNAL] replaced Saros class with proper preference stores
Replaced all access to Saros in the feedback package to use preference
stores.

Changed: tmp files are now created in the default tmp directory (who had
the clever idea to abuse the workspace ?!)

Open:

This patch does not really offer any benefit but it also does not
worthen things. At least it shows a nice deep insight into the Eclipse
preference mechanics which have to be abstracted for multi platform
usage.

Change-Id: I6b3b0d46a81c0cf86f791661a037ebfb967f7676
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1239
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>