[UI] Complement the SelectionAnnotation to a full highlighted block
When a remote user selected a text block, the highlighting on the
local screen looked rather ragged: only the *text* was highlighted, not
the white space unto the right margin.
This patch adds a new Annotation type that fills up this white space
with the same color as the character highlight.

For each session participant the original SelectionAnnotation and the
new FillUpAnnotation are drawn in same layer to prevent mixing up
annotations of different users (e.g. see plugin.xml).

Adapted from SWTP-2013 repository.
 - Authored by and reviewed by:
    Christoph Krüger <christoph.krueger@fu-berlin.de>
    Conrad Läßig <conrad.laessig@gmail.com>
    Richard Möhn <richard.moehn@fu-berlin.de>
 - Reviewed on (not a complete list)
    http://saros-build.imp.fu-berlin.de/gerrit/935
    http://saros-build.imp.fu-berlin.de/gerrit/1096
    http://saros-build.imp.fu-berlin.de/gerrit/1116

Change-Id: Ia8a5c879ce31ee6beb68857364cab5520324be97
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1236
Tested-by: Jenkins CI
Reviewed-by: Arndt Lasarzik <alasarzik@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>