[INTERNAL|REFACTOR] removed preferences from UPNPService
They were mainly used to remove previous port mappings.

Problem: The right way to remove those mappings is to query
the IGD for all port forwardings and then delete those that match.

As it is very uncommon that.
1. Eclipse crashes
2. A user or SMACK will change ports after the crash
3. The port mapping is still overwritten if it was created by Saros

It is removed and will be replaced later with the recommendation above

REFACTOR: moved most parts from the UPNPService to SarosNet because the
UPNPService should handle the port mapping and should not start to
retrieve the port from the Socks5 server.

Change-Id: I103298ee8d0b2fb3ca6a858f5002dd1e2e91f469
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1249
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>