[INTERNAL] removed StoppedAction
The action only exists (as stated by the Javadoc) for debugging
purposes.

On the other hand forcefully removing the blocking state does not offer
any benefit because either the logic already encountered a defect or if
a developer wish to cause inconsistencies there are multiple know
(easy) ways by hand so there is no reason for having such a switch.

Again ... if you wish to "debug" Saros you just have to read to trace
log. Successfully debugging Saros using the debugger is almost
impossible because it is not a single threaded application.

Change-Id: Ie6096e0380dfc2237067f4b1a3ae058a386e9cf9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1269
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>