[NOP] removed DiffPrint method
This method was used long ago when the watchdog detected inconsistencies
and the content to replaced was "diffed".

From time to time the diff produces errors so it was disabled.

Removed the logic as it is not offer any benefit because:

1. No developer will look at the output nor does it offer any
advantage because:

2. A "stupid" hex dump of the binary content (which is not done)
contains more detailed informations than the diff output.

Change-Id: I9fe731cf342d8b9459c01f42f0cbb21dd0b88b31
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1271
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>