[REFACTOR] moved static utility method
This method offered so MUCH utility that it was "ONLY" used in one
place.

I am getting sick of this utility overload ... time to put out -2 on
Gerrit for senseless utility class stuff.

Last but not least I love this article 
http://misko.hevery.com/2008/07/24/how-to-write-3v1l-untestable-code/
which nicely describes the issues if such "utility" methods do much more
than they are intended to do ! Again PowerMock is NOT the answer ... it
is not THE golden hammer against poor code design !!!

Change-Id: I1dc8493937dfe10f4b9290d704624270e92d8a95
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1273
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>