[INTERNAL] changed Version verification protocol
Up until know the packet that contains the version information contained
serialized Eclipse Objects and so this protocol would never be able to
communicate with other Saros applications that are running on other
platforms.

Changed the version packet, it now only contains string values as data
and so it is up to the remote side to correctly handle them which is now
at least possible.


Change-Id: Ie14cd99c47cf7594b71d9a7f6c7ca227c8006c95
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1277
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>