[INTERNAL] removed HTTP proxy settings for Smack
The Eclipse HTTP proxy settings are normally only set for accessing web
content. Since Smack needs a PERSISTENT HTTP connection is very unlikely
that the configured HTTP proxy will allow such connections and so the
user will never be able to connect to any XMPP server.

If HTTP support is still requested it should be reimplemented by adding
an extra field in the network preference page to allow the user to enter
such a proxy.

Change-Id: Ifa26a0f333ff58e1e9bd100f83bb9706d3afa5e3
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1298
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>