[BUILD] Created de.fu_berlin.inf.dpp.core project
This new project should contain all Eclipse-independent components of
Saros so that they can be used by other implementations (e.g. an
IntelliJ variant of Saros).
As a first, trivial step, this patch only moves the enumeration
"ConnectionState" to demonstrate the general idea.

One goal for the whole restructuring process is to keep the development
process simple, i.e. keep Saros "startable" directly from Eclipse (and,
if possible, other IDEs as well), without the need to create a separate
"saros_core.jar" for inclusion in the main Saros project, and so on.

However, the CI process needs a little adaptation: The Saros-Gerrit job
now looks for the core-project. If it's present, Jenkins uses the
build.xml files from the projects (which are also part of this patch);
otherwise, Jenkins behaves as before and uses the same build.xml for all
projects.
This double-configuration of the job will be removed as soon as there
are no more open patches without the core-project.

Change-Id: I86525767eb984e2d29965e7f5c2c0acca3bceeb1
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1309
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Arndt Lasarzik <alasarzik@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>