[FIX] fixed OOM error in DataTransferCollector
I do not know why the aggregation of the data was not done on the fly
but the current implementation had one major flow.

Assume ~= 50 byte per entry and so this would be a total of ~= 500 Kb
after 10000 transfer events. If the session would be run as sever it
should not take more than a month until JVM will throw an OOM.

Also fixed concurrency issues as two different threads read and write
the data.

Change-Id: I9e9b6d2a7ff129f1f69a7b75fed6f611f2733184
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1313
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>