[INTERNAL] removed SarosSession.MAX_USER variable
This variable may be highly confusing because it suggest that there is
an upper bound for the maximal users in a session.

Moved the "limit" variable to the SarosAnnotation class because this is
the only class that has currently a "limit" because every annotation
must be configured by hand in the plugin.xml file.

Change-Id: Ifdaf2f27587aeb96f0c0f3e0d5a89476aa63339a
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1329
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>