[API] removed *VOIP* feature
Removed the feature because it has several UX problems and last but not
least is very cumbersome. I personally tried it once and was very
disappointed. No noise reduction, no voice activation (you could here
everything in the background). Also this is not VOIP as VOIP uses RTP
(UDP packets) but this implementation uses plain TCP (sometimes tunneled
through XMPP) which may result in so much delay that it is not really
possible to start a communication. Last but not least it supports only 2
users which make DPartyP become DPairP once again.



Change-Id: I78e9c054888d916d24d6fd4cac09169dc283f356
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1340
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>