[INTERNAL] Stopping the ActivitySequencer will no longer stall forever
In case the remote side had gone offline due to system crash there is a
good change that the underlying transport will not detect this (e.g no
TCP RST packet received) and so stopping this component may take
forever.

This patch changes the behavior as it now only waits up to a timeout and
the just continue.

Also changed the start stop logic. The sequencer can no longer be
restarted but this should not matter as the SarosSession is garbage
collected anyway on session stop.

Change-Id: Ie02e85cf4af3907afd1c3bfb0986985df01efc32
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1383
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>