[INTERNAL] removed need based feature
This feature has several flaws.

See http://sourceforge.net/p/dpp/bugs/781/)

Beside the technical issues (mainly blocking everything through
modal dialogs) the usage of the feature is very poor. Either you can
enable it all the time, disable it for all the time or letting this
feature asking you over and over again.

The user should decide when to share a file, it should not be automated
as this may lead to bad results.


Change-Id: I1a5bd0d1213c2ed3b555c80501bbd3831b794bc7
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1409
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>