[GUI|API] allow the user to choose the Socks5 candidates
Depending on how much local network interfaces are present in the
current system the Socks5 connection establishment can really slow down
or result in timeouts although a Socks5 connection may be established.

Added an option for power users to change the behavior on how Saros
tries to establish Socks5 connections.

Change-Id: Ic5faaf4e3bebceaabe0a483d8a966a53b7a23b64
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1074
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>