[INTERNAL] removed JDT support
The JDT support was only used by the PreferenceManager class which
was removed with commit 6d9db93ef6329de5f33be2252516c3ecad5b5343.

Removed the whole support as it was a bad idea to manipulate project
settings without informing the user, so this should never happen again.

Last but not least this was only working for Java projects at all.

Change-Id: I756b2fe6cc6a93bf3fb179d9f4b62b92f8857837
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1414
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>