[FIX] Socks5Transport did not evaluate XMPP error codes
try
{
  ...
}
catch(/*Pokemon*/ Exception gotcha) {
   ... wait for remote side to connect
}

Well if the developer would have spend some time reading
the XEP-65 RFC it would be clear that the remote side may support Socks5
but do not accept the stream. This is currently the case when the "Use
IBB only" option is enabled.

The result of this Pokemon Exception handling: idle waiting (30 seconds)
for a remote connection request that will never been executed/received.

Change-Id: I53f55a0d15d10ca93cbb096d5936251635c628a5
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1417
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>