[INTERNAL] FileActitity util methods no longer read the file content
The file content must now be passed in. While working on the code I
figured out that the checksum was never used in the business logic.

Once again, the checksums used in the recovery by the Watchdog
Component are created from the EDITOR INPUT (as simple STRING !) and not
from the actual byte content of a given file.

Change-Id: I47dc20bab8a56286d8495930ffac7f8a829845cb
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1424
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>