[INTERNAL] moved User class into the core
Also changed the package name but using session instead of the old
project package.

Unfortunately this move breaks retrieving the nickname for a given user.

The affected !STATIC! methods were always questionable and this is
currently the result of such static calls. This will get fixed when the
missing class that those methods depends on is also moved into the core.



Change-Id: Ie7c37adb44b3093bf76c803937185e8f590a514d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1425
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>