[HACK] moved the creation of SharedProject into the SharedResourceManager
This is currently the easiest way to get rid of the SharedProject
reference in the ISarosSession interface. As the SharedProject class
also references VCS stuff which references Eclipse Subsclipse stuff this
is hardly to convert.

And again, this class is a dead left over when partial sharing was
introduced, only containing methods for logging and VCS stuff.


Change-Id: I359d2d72dbb395490a5eeedad0c3ec134665952d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1430
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>