[GUI] Show additional "cursor" in remote users' selections
Prior to this patch, the text cursors of remote users were shown as
one-character-wide selections. Now their cursors resemble the local
cursor (vertical, two-pixel-wide stroke).
This enables the local user to be sure whether the remote cursor is left
or right of the highlighted character.
To do so this commit adds an IDrawingStrategy.

This might also be the starting point for developing a name tag attached
to the cursor of each user.

Adapted from SWTP-2013 repository.
 - Commit: 2f38f4a
 - Reviewed on: http://saros-build.imp.fu-berlin.de/gerrit/893
 - Reviewed by: Richard Möhn <richard.moehn@fu-berlin.de>
 - Reviewed by: Conrad Läßig <conrad.laessig@gmail.com>

Change-Id: I0176465db072c447d543e31b5fc9f996c20114fa
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1234
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Tested-by: Jenkins CI