[INTERNAL] Do not refresh resources during FileList creation
Refreshing a resource may trigger resource change events that may have
an impact on the current session. Such a behavior is not desired when
creating file lists.

Refreshing of projects is now only performed once and that happens
BEFORE they are added to the session. This is to avoid that any resource
change events are broadcasted to the other session users.

Change-Id: I57133da7bbf3baeae1178aa7534ab7d5cd96ef39
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/751
Tested-by: Jenkins CI
Reviewed-by: Arsenij Solovjev <xeper000@gmail.com>
Reviewed-by: Damla Durmaz <ddurmaz88@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>