[REFACTOR] Rename ActivityProvider to ActivityProducerAndConsumer
The old name is confusing: currently Providers do both, provide and
consume.
And this is not only my opinion.

I've made the new name ridiculously long on purpose. This should serve
as a red flag indicating that something should be done about this. Also
the name does not lie about the nature of the interface.

After an optimal approach is found, I will make a followup patch.

Change-Id: I6844d1974dda0d03f7f42b743c2d0f8ba6fe302e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1466
Tested-by: Jenkins CI
Reviewed-by: Damla Durmaz <ddurmaz88@gmail.com>
Reviewed-by: Nils Bussas <nils.bussas@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>