[INTERNAL] Remove deprecated and unused method from ITransmitter
ITransmitter.sendToSessionUser(JID, PacketExtension) was deprecated in
favor of sendToSessionUser(String, JID, PacketExtension) in c36e99b0 (on
2013-09-13).
In the Saros plugin, the old method has not been used since.
The Saros whiteboard, however, still used the old method, so this commit
switches that to the new method and finally removes the old one.

Change-Id: Ibed8c1f88e054770485c03e222c1abc7ec245e2f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1470
Tested-by: Jenkins CI
Reviewed-by: Arndt Lasarzik <alasarzik@gmail.com>
Reviewed-by: Andrej Szaffranietz <a.szaffranietz@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>