[API] Revert 85d66af, introduce IActivityProducer and -Consumer instead
Thanks to Stefan Rossbach for pointing out a less intrusive way than
renaming the IActivityProvider interface to
IActivityProviderAndConsumer.

This patch deprecates the IActivityProvider interface and, in order to
take a first step towards the separation of concerns of Activity
producing and consuming, adds two new interfaces each of which is now
implemented by the AbstractActivityProvider class.

Change-Id: I5beaa2086aaeeedd5306687220d0ab87492c95da
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1477
Tested-by: Jenkins CI
Reviewed-by: Arsenij Solovjev <xeper000@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>