[FIX] Path-based editor closing was not working
The convenience methods EditorManager.isOpenEditor() and .closeEditor()
were not working because for finding the correct editor, they relied on
the equality of Eclipse's IResource and Saros Core's IResource (which is
always false).
This led to no compilation errors, because the equals() method is
declared on Object and takes any Object as parameter.
However, this defect did not lead to failures because (1) these methods
are only executed during inconsistency recovery which is seldom
necessary, and (2) even then, closing the affected editors does not seem
required to resolve the inconsistencies.

Change-Id: I56ce64f3b756146fd54dbc495042f07da855d743
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1497
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>