[REFACTOR] Full Revamp of the EnterProjectNamePage
Removes the hashmap nightmare. The options for a project are now handled
and displayed by a separate composite (ProjectOptionComposite).

This refactoring also includes some fixes regarding merging multiple
projects into a single one (which is not allowed).
Furthermore, the preselection of projects is done when necessary, i.e.
upon adding of new resources to a partially shared project. In
particular, projects are *not* preselected when they are not shared yet,
in order to avoid accidental overwrites.

Also, this patch cleans up the verification logic as it was almost
impossible to understand without spending several hours of reading the
source code again and again.

Change-Id: I830e6ae74bb3522391c73ae1eed6095f0514fa0e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1148
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>