[FIX] #823 Check Whiteboard availability during invitation
This patch prevents that a Whiteboard invitation process is even started
(or, as for the client, expected) when the remote side has no Whiteboard
plugin.
In sessions with more than two participants, the network part of the
host's Whiteboard will be initialized lazily (i.e. only when the first
client *with* a Whiteboard completed his session negotiation).
In general, in case the host has a Whiteboard, all participants that
have a Whiteboard too are able to use it.

The presence of the Whiteboard plugin is checked through an
InvitationHook.

Change-Id: Ia764bd375e54efb0ba7f63cdbfcbdca0d0dde3b2
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1506
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>