[API] Remove preIncomingInvitationCompleted()
The Whiteboard logic that relied on this listener was moved to
sessionStarted().

Furthermore, the progress monitor was removed from
SXEIncomingSynchronizationProcess. It was not accessible from the
GUI anyway, because the dialog with progress bar and Cancel
button is long gone when the client receives the SXE invitation.

Pair-programmed-with: Franz Zieris

Change-Id: I89c3b23a897f3357b2ade4e4ecc1b3cea85e5dfb
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1514
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>