[INTERNAL] ensure only resource qualified JIDs are used for a user
Just added a small check to ensure that resource qualified JIDs are used
for a user during a session. As this was already the case this patch
does not change the logic at all.

For further development (e.g different transport layer like plain TCP
instead of XMPP) the getResourceQualifiedJID method of the ISarosSession
interface is now deprecated.


Change-Id: I12865dbf80d62023248c816c71fe01dc43c06e48
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1551
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>