[FIX] GUI condition "sessionHeadlineSelected" referenced wrong class
This bug was introduced in 484855a, probably due to an incomplete
refactoring, i.e. moving of SessionHeaderElement from package
"rosterSession" to "session" which was not applied to the plugin.xml.

This led to the failure of two STF test cases. They could not find the
expected entries in the context menu, which was not displayed because
the condition to display that menu referred to the wrong class name, so
it could never hold.

Normally, the Eclipse refactoring methods should cover such
references in plug-in manifest files. 

Change-Id: I9d2f7190bbf480d08988c32fc7662a8858f2721b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1570
Tested-by: Jenkins CI
Reviewed-by: Damla Durmaz <ddurmaz88@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>