[INTERNAL|API] extended IFACE to access runtime components
It is now possible to access components that reside in the session
context.

Deprecated the new method per default as it is a dangerous call but
currently allows an easy access for GUI component to retrieve such
components without extending existing listeners that a currently
available to the GUI classes.


Change-Id: I1b012c30014ee4d4be2b6f783d00623817c08cec
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1585
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>