[INTERNAL] moved connection logic out of Saros class
There exists no multi inheritance in Java so it is only possible
to use interface e.g ISaros which would result in multiple
implementations. As no developer wants to fix bugs in multiple
implementations which are mostly the result of a C&P and small
modifications those logic is extracted out of the Saros class to a small
facade class.


Change-Id: Id202c19baa936ccd380ef3e33f834b42f967a5a7
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1301
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>