[INTERNAL] XMPPAccountStore is now platform independent
The encryption is now done by the store itself. This should be at least
as secure as using OSGi SecuredPreferences if the account file is stored
in the same place where the OSGi framework would store the key file.

If you really want to obtain the data this is also possible with OSGi if
you are able to access the password file as the whole encryption
mechanism is also open source code.

Change-Id: I6305ca1f035caa91e656f3e45fbca7a90ddbab0d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1422
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>