[INTERNAL] removed getUserName* from User class.
Those display name were tightly coupled with XMPP. Introduced a new
nickname variable that can now safely used to display the name of a user
during a session at it is independent of the network layer that this
user uses.

Currently there is NO option to either choose the own nickname nor is
this nickname transmitted to other session users. Those logic will be
part of another patches.

For now the nickname of the user is its XMPP address (JID) without
resource qualifier.

Change-Id: I92cc96149dcfe66aaa92dd06f473fdd2025cad1c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1612
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>