[FEATURE] user can now choose their own nicknames for a session
Users can now use a custom nickname. This nickname will be visible to
all session users.

This may a bit confusing for now as the session chat is a MUC that
is hosted on a XMPP server and therefore the nicknames of the XMPP
entries are use which may now be completely different.

Change-Id: I3b088fd27cee4bdbafc841a6202698d2eb8bc9b3
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1619
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>