[INTERNAL] undo i18n for plugin.xml
Inlined back all strings. Also cleaned up some commands.

TODO: Even as an experienced Saros developers I find the GUI menu
structure very complex. There are so much options on how to start
sharing. If looking at the feedback it seems that the overall option to
start a project sharing is to right click on a contact in the contact
list and then choose the project to share and so missing at least the
partial sharing feature. This feels wrong because at least NO hint is
displayed that the project sharing offers additional setup steps.

At least we can think about to remove the session / sharing stuff from
the Saros Menu entry in the Eclipse menu bar.

Change-Id: I59038747672d75cfd768e7c2d13503bc17f1ac49
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1634
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>