[INTERNAL] fix for commit 783dec3b76ef71f2a6d9e9bf097139401118ba26
Due the transparent class loading mechanism in Java the commit
introduced an IllegalAccessError when trying to access the SHARED_LOCK
variable.

This field is initialized by the classloader that loads the Saros Core
files. Unluckily the OSN/ISN classes are currently loaded by the
classloader for the Saros Eclipse Plugin.

See
http://stackoverflow.com/questions/10465646/cant-access-protected-member-of-superclass-from-same-package-in-different-jar

As the static variable is NOT inherited it cannot be accessed ...

TLDR

OSN tries to access SHARED_LOCK (but not the SHARED_LOCK that was loaded
with the class loader of OSN) but the SHARED_LOCK of CancelableProcess
(loaded with class loader of CancelableProcess). This is forbidden. You
cannot access static fields of a class that is loaded with a different
class loader.

Change-Id: Ia528b282c9b576e6301517b907a0dd0e94a13753
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1660
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>