[INTERNAL] removed option to display non Saros Contacts
The FilteredContactSelection composite allowed to display contacts that
do not support Saros.

Due to the bad implementation updates of users that goes online during
this the presentation are not correctly updated.

I do not really know what the intention of the original developer was
but maybe he/she should just add "Refresh" button instead of this
behavior.


Change-Id: Idda15659296017dc29160c997d56cb8c5d3b4608
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1726
Tested-by: Jenkins-Eclipse QA
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>