[FEATURE] Action Awareness: Show open dialogs/active views of others
This feature tracks opened dialogs or activated views of session
participants and sends the title of the opened dialog or activated view
to the other participants of the session.

A new DialogAwarenessTreeElement is appended to the user element with a
nice "file object" icon, and the title of the opened dialog or activated
view, so other users see: Ah, Bob has the "New Java Class" dialog open
or: Ah, Alice is in the "Tasks" view.

Additionally, the awareness information is printed in the multi user
chat and in the status line of the Eclipse workbench. This is done for
further user experiments to investigate, which place is the preferred
one for action awareness information. This can be configured in the
saros-properties file. Currently, just the roster is configured to
display the awareness information.

This feature is based on patch #52 by Alexander Waldmann, which was not
submitted:
http://saros-build.imp.fu-berlin.de/gerrit/#/c/52/

Change-Id: Iff6803128614d785f62a528646f3467c70505ad2
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1572
Tested-by: Saros-Core QA
Tested-by: Saros-Eclipse QA
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>