[FEATURE] Action Awareness: Show test runs with their result
This feature tracks test runs of session participants and sends the
title and the result of started and finished test runs to the other
participants of the session.

The information is displayed in the new DialogAwarenessTreeElement,
introduced by the new action awareness feature, see:
http://saros-build.imp.fu-berlin.de/gerrit/#/c/1572/

Additionally, the awareness information is printed in the multi user
chat and in the status line of the Eclipse workbench. This is done for
further user experiments to investigate, which place is the preferred
one for action awareness information. This can be configured in the
saros-properties file. Currently, just the roster is configured to
display the awareness information.

Change-Id: Ifff8311bbfc35839c3dc6d15418a4daba7dc3fb5
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1788
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>